# specs/core-gateway.openapi.yaml
openapi: 3.1.0
info:
  title: StarlingAI Gateway API
  version: 0.7.0
  summary: HTTP + WebSocket surface for the StarlingAI agent-swarm gateway.
  description: |
    REST API for the `@starlingai/core` gateway. Companion WebSocket RPC is at
    `ws://${host}:${port}/ws?token=<jwt>` and follows JSON-RPC 2.0 (see
    `docs/api.md` for the RPC method catalog).

    Authentication: all `/api/*` paths require a bearer JWT unless noted.
    Trigger endpoints also accept a webhook key via `?key=` or a header — but the
    header name differs per endpoint: `POST /api/scenes/{name}/run` reads
    `X-Scene-Key`, while `POST /api/jobs/{name}/run` reads `X-Job-Key`.

    Schemas below are intentionally loose — the runtime validates with Zod
    schemas in `packages/core/src/config/` and `packages/core/src/gateway/`.
    Refer to those schemas for exhaustive field definitions.
  license:
    name: CC BY-NC 4.0
    url: https://creativecommons.org/licenses/by-nc/4.0/
servers:
  - url: http://localhost:8765
    description: Local default (gateway.port)
  - url: https://{host}
    description: Reverse-proxied deployment
    variables:
      host:
        default: ai.internal.example
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sceneKey:
      type: apiKey
      in: header
      name: X-Scene-Key
    jobKey:
      type: apiKey
      in: header
      name: X-Job-Key
  schemas:
...
    PersonalityConfig:
      type: object
      properties:
        systemPrompt: { type: string }
        source: { type: string, enum: [config, runtime] }
...
security:
  - bearerAuth: []
paths:
...
  /api/agents/resolve:
    get:
      tags: [Agents]
      summary: Resolve a query to the best-matched agent(s)
      parameters:
        - { name: query, in: query, required: true, schema: { type: string }, description: 'Text to route. Missing/empty returns 400.' }
        - { name: minConfidence, in: query, required: false, schema: { type: string, enum: [high, medium, low], default: medium } }
      responses:
        '200': { description: OK }
        '400': { description: '`query` missing or empty' }
  /api/agents/{name}/model:
    patch:
      tags: [Agents]
      summary: Set an agent's model config, saved to the config overlay
      parameters: [ { name: name, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses: { '200': { description: Saved }, '400': { description: 'Invalid value, or a saved key would go to a new endpoint' }, '404': { description: Unknown agent }, '409': { description: 'Agent is not in the saved config (bridged from an A2A peer)' } }
...
  /api/flow-memory:
    get: { tags: [Memory], summary: Read persistent swarm flow memory, responses: { '200': { description: OK } } }
    post:
      tags: [Memory]
      summary: Append to flow memory
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses: { '200': { description: Appended } }
  /api/config-assistant/proposals:
    get: { tags: [ConfigAssistant], summary: List pending config-assistant proposals, responses: { '200': { description: OK } } }
    post:
      tags: [ConfigAssistant]
      summary: Create a config-assistant proposal
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses: { '201': { description: Created } }
...
  /api/personality:
    get: { tags: [Personality], summary: Read the main assistant system prompt, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PersonalityConfig' } } } } } }
    put:
      tags: [Personality]
      summary: Replace the main assistant system prompt
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { systemPrompt: { type: string } }, required: [systemPrompt] } } } }
      responses: { '200': { description: Updated } }
...